requirements.txt
playwright>=1.52,<2
pandas>=2.2,<3
python-dotenv>=1.0,<2
gspread>=6.1,<7
google-auth>=2.29,<3
